# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateServiceRequest:
    properties:
      description:
        type: string
      icon:
        type: string
      image:
        type: string
      is_active:
        type: boolean
      slug:
        type: string
      sort_order:
        type: integer
      title:
        type: string
    required:
    - description
    - title
    type: object
  dto.CreateTestimonialRequest:
    properties:
      company:
        type: string
      content:
        type: string
      designation:
        type: string
      image:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      rating:
        type: integer
      sort_order:
        type: integer
    required:
    - content
    - name
    type: object
...
  dto.UpdateServiceRequest:
    properties:
      description:
        type: string
      icon:
        type: string
      image:
        type: string
      is_active:
        type: boolean
      slug:
        type: string
      sort_order:
        type: integer
      title:
        type: string
    type: object
  dto.UpdateTestimonialRequest:
    properties:
      company:
        type: string
      content:
        type: string
      designation:
        type: string
      image:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      rating:
        type: integer
      sort_order:
        type: integer
    type: object
...
  models.Service:
    properties:
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      image:
        type: string
      is_active:
        type: boolean
      slug:
        type: string
      sort_order:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.Testimonial:
    properties:
      company:
        type: string
      content:
        type: string
      created_at:
        type: string
      designation:
        type: string
      id:
        type: integer
      image:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      rating:
        type: integer
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  utils.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      success:
        type: boolean
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API server for GK Capital finance website.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GK Capital API Documentation
  version: "1.0"
paths:
...
  /api/admin/services:
    get:
      description: Get paginated list of all services including inactive ones
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all services for administration
      tags:
      - Admin Services
    post:
      consumes:
      - application/json
      description: Create a service for administration
      parameters:
      - description: Create Service Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateServiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Service'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new service
      tags:
      - Admin Services
  /api/admin/services/{id}:
    delete:
      description: Delete a service by ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a service
      tags:
      - Admin Services
    put:
      consumes:
      - application/json
      description: Update details of a service by ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Service Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Service'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update an existing service
      tags:
      - Admin Services
  /api/admin/testimonials:
    get:
      description: Get paginated list of all testimonials including inactive ones
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Get all testimonials for administration
      tags:
      - Admin Testimonials
    post:
      consumes:
      - application/json
      description: Create a testimonial for administration
      parameters:
      - description: Create Testimonial Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTestimonialRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Testimonial'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new testimonial
      tags:
      - Admin Testimonials
  /api/admin/testimonials/{id}:
    delete:
      description: Delete a testimonial by ID
      parameters:
      - description: Testimonial ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a testimonial
      tags:
      - Admin Testimonials
    put:
      consumes:
      - application/json
      description: Update details of a testimonial by ID
      parameters:
      - description: Testimonial ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Testimonial Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTestimonialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Testimonial'
              type: object
      security:
      - BearerAuth: []
      summary: Update an existing testimonial
      tags:
      - Admin Testimonials
...
  /api/services:
    get:
      description: Get list of all active services for public view
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Service'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Get active services
      tags:
      - Public Services
  /api/services/{id}:
    get:
      description: Get details of a single service by ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Service'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Get service by ID
      tags:
      - Public Services
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"